refactor: models/cart.ts로 비즈니스 로직 이동
